unify read test code in common
